fix(hud): extract useHudState to own file, fix lint errors
