+++ frontend/src/components/GameBoard/hud/GameHud.tsx
@@ -1,39 +1,10 @@
-import { useState, useEffect, useRef } from 'react';
 import type { RefObject } from 'react';
 import type { GameStateSnapshot } from '@/game/types';
 import { CharacterState } from '@/game/constants';
-import { extractHudState, hudStateEqual, type HudState } from './types';
+import { useHudState } from './useHudState';
 import ResourceBar from './ResourceBar';
 import AbilityBar from './AbilityBar';
 import './hud.css';
-
-export function useHudState(
-	snapshotRef: RefObject<GameStateSnapshot | null>,
-	localPlayerId: number,
-): HudState | null {
-	const [hud, setHud] = useState<HudState | null>(null);
-	const lastUpdate = useRef(0);
-
-	useEffect(() => {
-		let raf: number;
-		const poll = (now: number) => {
-			if (now - lastUpdate.current >= 66) {
-				const snap = snapshotRef.current;
-				const char = snap?.characters.find((c) => c.player_id === localPlayerId);
-				if (char) {
-					const next = extractHudState(char);
-					setHud((prev) => (prev && hudStateEqual(prev, next) ? prev : next));
-				}
-				lastUpdate.current = now;
-			}
-			raf = requestAnimationFrame(poll);
-		};
-		raf = requestAnimationFrame(poll);
-		return () => cancelAnimationFrame(raf);
-	}, [snapshotRef, localPlayerId]);
-
-	return hud;
-}
 
 interface GameHudProps {
 	snapshotRef: RefObject<GameStateSnapshot | null>;
